requirements.txt
playwright==1.51.0
pytest==8.3.5
python-dotenv==1.1.0
faker==37.1.0